@@ -1,6 +1,6 @@
 module web
 
-go 1.18
+go 1.21
 
 require (
 	github.com/mattn/go-sqlite3 v1.14.14
